requirements.txt
Flask==2.3.3
Werkzeug==2.3.7
requests==2.31.0
boto3==1.34.0
botocore==1.34.0
gunicorn==21.2.0
python-dotenv==1.0.0
bcrypt==4.0.1
PyJWT==2.8.0
